+++ apm-sniffer/apm-sdk-plugin/grpc-1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/grpc/v1/server/TracingServerCall.java
@@ -92,7 +92,7 @@ public class TracingServerCall<REQUEST, RESPONSE> extends ForwardingServerCall.S
                 }
                 break;
         }
-        Tags.RESPONSE_CODE.set(span, status.getCode().name());
+        Tags.RPC_RESPONSE_STATUS_CODE.set(span, status.getCode().name());
 
         try {
             super.close(status, trailers);