+++ apm-sniffer/apm-sdk-plugin/jsonrpc4j-1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/jsonrpc4j/JsonServiceExporterInterceptor.java
@@ -68,7 +68,7 @@ public class JsonServiceExporterInterceptor implements InstanceMethodsAroundInte
         HttpServletResponse response = (HttpServletResponse) objects[1];
         AbstractSpan span = ContextManager.activeSpan();
         if (IS_SERVLET_GET_STATUS_METHOD_EXIST) {
-            Tags.STATUS_CODE.set(span, response.getStatus());
+            Tags.HTTP_RESPONSE_STATUS_CODE.set(span, response.getStatus());
         }
 
         ContextManager.stopSpan();