+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-2.0.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v20x/HttpClientRequestInterceptor.java
@@ -92,7 +92,7 @@ public class HttpClientRequestInterceptor implements InstanceMethodsAroundInterc
                     } else if (httpClientResponse.status().code() > 400) {
                         abstractSpan.errorOccurred();
                     }
-                    Tags.STATUS_CODE.set(abstractSpan, httpClientResponse.status().code());
+                    Tags.HTTP_RESPONSE_STATUS_CODE.set(abstractSpan, httpClientResponse.status().code());
                     abstractSpan.asyncFinish();
                 }