+++ oap-server/analyzer/agent-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/provider/trace/parser/SpanTags.java
@@ -22,9 +22,9 @@ package org.apache.skywalking.oap.server.analyzer.provider.trace.parser;
  * Reserved keys of the span. The backend analysis the metrics according the existed tags.
  */
 public class SpanTags {
-    public static final String STATUS_CODE = "status_code";
+    public static final String HTTP_RESPONSE_STATUS_CODE = "http.status_code";
 
-    public static final String RESPONSE_CODE = "response_code";
+    public static final String RPC_RESPONSE_STATUS_CODE = "rpc.status_code";
 
     public static final String DB_STATEMENT = "db.statement";