+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/All.java
@@ -53,10 +53,14 @@ public class All extends Source {
     private boolean status;
     @Getter
     @Setter
+    @Deprecated
     private int responseCode;
     @Getter
     @Setter
-    private String responseCodeStr;
+    private int httpResponseStatusCode;
+    @Getter
+    @Setter
+    private String rpcStatusCode;
     @Getter
     @Setter
     private RequestType type;